3.5 переделал на интерфейсы и убрал константу
